Rebased and changed drive.h
